% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{add_app_dir}
\alias{add_app_dir}
\title{Add a directory for you app}
\usage{
add_app_dir()
}
\description{
This function creates a directory called \code{app} in your project.
All your files and directories created with \code{{box.shiny}} will be placed inside
of this directory.
}
