% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{add_app_files}
\alias{add_app_files}
\title{Add main application files}
\usage{
add_app_files()
}
\description{
Creates files \code{app/app_ui.R}, \code{app/app_server.R}, \code{app/run_app.R} and \code{app.R}
(if they don't already exist).
\itemize{
\item \code{app/app_ui.R}: The application UI function
\item \code{app/app_server.R}: The application server function
\item \code{app/run_app.R}: The function to run the application
\item \code{app.R}: File for running the application
}
}
