% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_helpers.R
\name{add_dir_helpers}
\alias{add_dir_helpers}
\title{Add directories for helper functions}
\usage{
add_dir_helpers()
}
\description{
Creates the following directories:
\itemize{
\item \code{app/helpers}
\item \code{app/helpers/ui}
\item \code{app/helpers/server}
}
}
