% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_helpers.R
\name{add_helper_server}
\alias{add_helper_server}
\title{Add files for server helper functions}
\usage{
add_helper_server(helper_name, open = TRUE)
}
\arguments{
\item{helper_name}{Name of the server helper (single character value)}

\item{open}{Whether to open the file or not}
}
\description{
Add files for server helper functions
}
