% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_helpers.R
\name{add_helper_ui}
\alias{add_helper_ui}
\title{Add files for UI helper functions}
\usage{
add_helper_ui(name, open = TRUE)
}
\arguments{
\item{open}{(\code{logical(1)})\cr
Whether to open the file or not.}

\item{helper}{(\code{character(1)})\cr
Name of the UI helper.}
}
\description{
Add files for UI helper functions
}
