% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_main_dir}
\alias{add_main_dir}
\title{Create main module directory}
\usage{
add_main_dir(dir_name = "shinymodules", pkg = TRUE)
}
\arguments{
\item{dir_name}{The name of your main module directory}

\item{pkg}{whether you are developing your app as a package.
If TRUE then the directory will be created inside of the `/inst` directory.}
}
\description{
This function is intended to create the directory for your modules.
}
