% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{add_main_module_dir}
\alias{add_main_module_dir}
\title{Create main module directory}
\usage{
add_main_module_dir(...)
}
\arguments{
\item{...}{Single character values to construct the path to your main module directory
(see \code{fs::path}).}
}
\description{
This function is intended to create the top level directory for your modules inside of
your \code{app} directory.
On the one hand it will create the main module directory. On the other hand, it will also
create a file \code{shinymodules_main_dir.txt} in your working directory, which will contain the
relative path to the main module directory which is used by other functions of the \code{{box.shiny}}
package.
}
