% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_modules.R
\name{add_module_dir}
\alias{add_module_dir}
\title{Create a module directory}
\usage{
add_module_dir(...)
}
\arguments{
\item{...}{(\code{character(1)})\cr
An arbitrary number of strings used to construct the path to the module directory
(see \code{fs::path()}).}
}
\description{
Create a module directory
}
