% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_modules.R
\name{add_module_file}
\alias{add_module_file}
\title{Create a module file}
\usage{
add_module_file(
  ...,
  ui_params = NULL,
  server_params = NULL,
  main = FALSE,
  open = TRUE
)
}
\arguments{
\item{...}{(\code{character(1)})\cr
An arbitrary number of strings used to construct the path to the module file
(see \code{fs::path()}).}

\item{ui_params}{(\code{character()})\cr
A character vector of additional UI arguments you wish to include.}

\item{server_params}{(\code{character()})\cr
A character vector of additional server arguments you wish to include.}

\item{main}{(\code{logical(1)})\cr
Indicate whether this is your main / top level module.
The main module is one, which will be placed directly into the UI and server
functions of the app.}

\item{open}{(\code{logical(1)})\cr
Whether to open the file or not.}
}
\description{
Create a module file
}
