% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_R6Class.R
\name{add_R6Class}
\alias{add_R6Class}
\title{Create a file for a new R6 class generator}
\usage{
add_R6Class(class_name, open = TRUE)
}
\arguments{
\item{class_name}{The name of the R6 class.}

\item{open}{Whether to open the file or not.}
}
\description{
Create a file for a new R6 class generator
}
