% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sass_partial.R
\name{add_sass_partial}
\alias{add_sass_partial}
\title{Add a \code{.scss} file}
\usage{
add_sass_partial(name, main_import = TRUE, open = TRUE)
}
\arguments{
\item{name}{(\code{character(1)})\cr
A name identifying the .scss file.}

\item{main_import}{(\code{logical(1)})\cr
Whether to add a corresponding import declaration to the main .scss file.}

\item{open}{(\code{logical(1)})\cr
Whether to open the file after it has been created.}
}
\description{
Add a \code{.scss} file
}
