% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-setup.R
\name{setup_project}
\alias{setup_project}
\title{Set up the infrastructure for the shiny app}
\usage{
setup_project(dir = getwd())
}
\arguments{
\item{dir}{(\code{character(1)})\cr
The directory in which to create the project.}
}
\description{
Set up the infrastructure for the shiny app
}
