% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_helper.R
\name{use_helper_ui}
\alias{use_helper_ui}
\title{Add predefined UI helper functions from the \code{{box.shiny}} package to your app}
\usage{
use_helper_ui(choice)
}
\arguments{
\item{choice}{(\code{character()})\cr
A selection of which helper functions to add.}
}
\description{
Currently available choices for helper functions are:
\itemize{
\item "columns": Wrappers around shiny's \code{column} function
}
}
