% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_helpers.R
\name{add_helper_server}
\alias{add_helper_server}
\title{Add files for server helper functions}
\usage{
add_helper_server(name, open = TRUE)
}
\arguments{
\item{name}{(\code{character(1)})\cr
Name of the server helper.}

\item{open}{(\code{logical(1)})\cr
Whether to open the file or not.}
}
\description{
Add files for server helper functions
}
