% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_helpers.R
\name{add_helper_ui}
\alias{add_helper_ui}
\title{Add files for UI helper functions}
\usage{
add_helper_ui(helper_name, open = TRUE)
}
\arguments{
\item{helper_name}{Name of the UI helper (single character value)}

\item{open}{Whether to open the file or not}
}
\description{
Add files for UI helper functions
}
