% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_funs.R
\name{add_module_dir}
\alias{add_module_dir}
\title{Create a module directory}
\usage{
add_module_dir(mod_name, main_dir = "shinymodules", pkg = TRUE)
}
\arguments{
\item{mod_name}{The name/path of the new module directory}

\item{main_dir}{The name of the main module directory}

\item{pkg}{Wether you are developing your app as a package.
If TRUE then the directory will be created inside of the `/inst` directory.}
}
\description{
Create a module directory
}
\examples{
add_module_dir("tabs")
add_module_dir("tabs/tab_one")
}
