% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_modules.R
\name{add_module_file}
\alias{add_module_file}
\title{Create a module file}
\usage{
add_module_file(
  ...,
  ui_params = NULL,
  server_params = NULL,
  main = FALSE,
  open = TRUE
)
}
\arguments{
\item{...}{Single character values to construct the path to your module file
(see \code{fs::path}, but \code{ext} argument can and must be ignored).
The file will be created underneath the main module directory created with \code{box.shiny::add_main_dir}.}

\item{ui_params}{Character vector of additional UI arguments you wish to include.}

\item{server_params}{Character vector of additional server arguments you wish to include.}

\item{main}{Logical value to indicate if this is your main / top level module.
This is a module that is going to be placed directly into the UI and server functions
of your shiny app.}

\item{open}{Logical, whether to open the file or not.}
}
\description{
This function creates an R-file with your shiny module template. The argument \code{mod_name} needs to be the
path of your module file, however you don't need to (and shouldn't) add the ".R" file extension to it.
}
