% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_R6Class.R
\name{add_R6Class}
\alias{add_R6Class}
\title{Create a file for a new R6 class generator}
\usage{
add_R6Class(name, open = TRUE)
}
\arguments{
\item{name}{(\code{character(1)})\cr
The name of the R6 class.}

\item{open}{(\code{logical(1)})\cr
Whether to open the file or not.}
}
\description{
Create a file for a new R6 class generator
}
