% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sass_partial.R
\name{add_sass_partial}
\alias{add_sass_partial}
\title{Add a \code{.scss} file}
\usage{
add_sass_partial(name, main_import = TRUE, open = TRUE)
}
\arguments{
\item{name}{A name identifying the .scss file.}

\item{main_import}{Whether to add a corresponding import declaration to the
main .scss file.}

\item{Whether}{to open the file after it has beeen created.}
}
\description{
Add a \code{.scss} file
}
