% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-setup.R
\name{setup_project}
\alias{setup_project}
\title{Set up the infrastructure for the shiny app}
\usage{
setup_project(module_dir)
}
\arguments{
\item{module_dir}{...}
}
\description{
Set up the infrastructure for the shiny app
}
